requirements.txt
Flask==3.0.0
pytest==7.4.0
pytest
google-cloud-aiplatform
ipytest
mock
pytest-mock
google-auth